pyproject.toml
[tool.poetry]
name = "django-celery"
version = "0.1.0"
description = ""
authors = ["muhtasim1 <muhtasim@greentomatomedia.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangoadmin = "^2.2.8"
djangorestframework = "^3.13.1"
celery = "^5.2.7"
redis = "^4.3.1"
flower = "^1.0.0"
django-celery-beat = "^2.2.1"
psycopg2-binary = "^2.9.3"
django-debug-toolbar = "^3.7.0"
django-rest-swagger = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
